Restructuration of xclim
