+++ docs/api.rst
@@ -1,23 +1,69 @@
+===
 API
 ===
 
 Indices
--------
+=======
 
 .. toctree::
 
    indices
 
-Submodules
-----------
+Indices submodules
+------------------
 
-checks module
--------------
-
-.. automodule:: xclim.checks
+.. automodule:: xclim.indices.calendar
    :members:
    :undoc-members:
    :show-inheritance:
+
+.. automodule:: xclim.indices.generic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: xclim.indices.run_length
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: xclim.indices.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Unit handling module
+====================
+
+.. automodule:: xclim.units
+   :members:
+   :undoc-members:
+   :imported-members:
+   :show-inheritance:
+
+
+Indicators
+==========
+
+.. automodule:: xclim.indicators.checks
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: xclim.indicators.indicator
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: xclim.indicators.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Other xclim modules
+===================
 
 ensembles module
 ----------------
@@ -27,35 +73,11 @@ ensembles module
    :undoc-members:
    :show-inheritance:
 
-generic module
---------------
-
-.. automodule:: xclim.generic
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-run\_length module
-------------------
-
-.. automodule:: xclim.run_length
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 
 subset module
 -------------
 
 .. automodule:: xclim.subset
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-utils module
-------------
-
-.. automodule:: xclim.utils
    :members:
    :undoc-members:
    :show-inheritance:
