+++ docs/index.rst
@@ -19,9 +19,10 @@ xclim currently provides over 50 indices related to mean, minimum and maximum da
    notebooks/usage
    notebooks/index
    indicators
-   api
    notebooks/units
+   internationalization
    icclim
+   api
    contributing
    authors
    history
