+++ docs/internationalization.rst
@@ -1,0 +1,3 @@
+.. automodule:: xclim.locales
+   :members:
+   :show-inheritance:
