+++ tests/conftest.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pytest
 import xarray as xr
 
-import xclim.utils as utils
+from xclim.indices.calendar import calendars
 
 
 @pytest.fixture
@@ -144,7 +144,7 @@ def ndq_series():
 @pytest.fixture
 def per_doy():
     def _per_doy(values, calendar="standard", units="kg m-2 s-1"):
-        n = utils.calendars[calendar]
+        n = calendars[calendar]
         if len(values) != n:
             raise ValueError(
                 "Values must be same length as number of days in calendar."
