+++ tests/test_checks.py
@@ -5,8 +5,8 @@ import pandas as pd
 import pytest
 import xarray as xr
 
-from xclim import checks
-from xclim.atmos import tg_mean
+from xclim.indicators import checks
+from xclim.indicators.atmos import tg_mean
 
 K2C = 273.15
