+++ tests/test_land.py
@@ -1,38 +1,35 @@
 import numpy as np
 import pytest
 
-with pytest.warns(DeprecationWarning):
-    from xclim import streamflow
+from xclim import land
 
 
 def test_base_flow_index(ndq_series):
-    out = streamflow.base_flow_index(ndq_series, freq="YS")
+    out = land.base_flow_index(ndq_series, freq="YS")
     assert out.attrs["units"] == ""
 
 
 class Test_FA:
     def test_simple(self, ndq_series):
-        out = streamflow.freq_analysis(
+        out = land.freq_analysis(
             ndq_series, mode="max", t=[2, 5], dist="gamma", season="DJF"
         )
         assert out.long_name == "N-year return period max winter 1-day flow"
         assert out.shape == (2, 2, 3)  # nrt, nx, ny
 
     def test_no_indexer(self, ndq_series):
-        out = streamflow.freq_analysis(ndq_series, mode="max", t=[2, 5], dist="gamma")
+        out = land.freq_analysis(ndq_series, mode="max", t=[2, 5], dist="gamma")
         assert out.long_name == "N-year return period max annual 1-day flow"
         assert out.shape == (2, 2, 3)  # nrt, nx, ny
 
     def test_q27(self, ndq_series):
-        out = streamflow.freq_analysis(
-            ndq_series, mode="max", t=2, dist="gamma", window=7
-        )
+        out = land.freq_analysis(ndq_series, mode="max", t=2, dist="gamma", window=7)
         assert out.shape == (1, 2, 3)
 
     def test_empty(self, ndq_series):
         q = ndq_series.copy()
         q[:, 0, 0] = np.nan
-        out = streamflow.freq_analysis(
+        out = land.freq_analysis(
             q, mode="max", t=2, dist="genextreme", window=6, freq="YS"
         )
         assert np.isnan(out.values[:, 0, 0]).all()
@@ -40,14 +37,14 @@ class Test_FA:
 
 class TestStats:
     def test_simple(self, ndq_series):
-        out = streamflow.stats(ndq_series, freq="YS", op="min", season="MAM")
+        out = land.stats(ndq_series, freq="YS", op="min", season="MAM")
         assert out.attrs["units"] == "m^3 s-1"
 
     def test_missing(self, ndq_series):
         a = ndq_series
         a = ndq_series.where(~((a.time.dt.dayofyear == 5) * (a.time.dt.year == 1902)))
         assert a.shape == (5000, 2, 3)
-        out = streamflow.stats(a, op="max", month=1)
+        out = land.stats(a, op="max", month=1)
 
         np.testing.assert_array_equal(out.sel(time="1900").isnull(), False)
         np.testing.assert_array_equal(out.sel(time="1902").isnull(), True)
@@ -55,16 +52,16 @@ class TestStats:
 
 class TestFit:
     def test_simple(self, ndq_series):
-        ts = streamflow.stats(ndq_series, freq="YS", op="max")
-        p = streamflow.fit(ts, dist="gumbel_r")
+        ts = land.stats(ndq_series, freq="YS", op="max")
+        p = land.fit(ts, dist="gumbel_r")
         assert p.attrs["estimator"] == "Maximum likelihood"
 
 
 def test_qdoy_max(ndq_series, q_series):
-    out = streamflow.doy_qmax(ndq_series, freq="YS", season="JJA")
+    out = land.doy_qmax(ndq_series, freq="YS", season="JJA")
     assert out.attrs["units"] == ""
 
     a = np.ones(450)
     a[100] = 2
-    out = streamflow.doy_qmax(q_series(a), freq="YS")
+    out = land.doy_qmax(q_series(a), freq="YS")
     assert out[0] == 101
