+++ xclim/indicators/seaIce/_seaice.py
@@ -4,9 +4,9 @@ Sea ice indicators
 """
 import abc
 
-from xclim import checks
 from xclim import indices
-from xclim.indicator import Indicator2D
+from xclim.indicators import checks
+from xclim.indicators import Indicator2D
 
 __all__ = ["sea_ice_area", "sea_ice_extent"]
