+++ xclim/indices/_threshold.py
@@ -1,13 +1,14 @@
 import datetime
-from typing import Union
 
 import numpy as np
 import xarray
 
-from xclim import run_length as rl
-from xclim import utils
-from xclim.utils import declare_units
-from xclim.utils import units
+from xclim.indices import run_length as rl
+from xclim.indices.generic import threshold_count
+from xclim.units import convert_units_to
+from xclim.units import declare_units
+from xclim.units import pint_multiply
+from xclim.units import units
 
 xarray.set_options(enable_cftimeindex=True)  # Set xarray to use cftimeindex
 
@@ -76,7 +77,7 @@ def cold_spell_days(tas, thresh="-10 degC", window: int = 5, freq="AS-JUL"):
     where :math:`[P]` is 1 if :math:`P` is true, and 0 if false.
 
     """
-    t = utils.convert_units_to(thresh, tas)
+    t = convert_units_to(thresh, tas)
     over = tas < t
     group = over.resample(time=freq)
 
@@ -126,7 +127,7 @@ def daily_pr_intensity(pr, thresh="1 mm/day", freq="YS"):
     >>> pr = xr.open_dataset("pr_day.nc").pr
     >>> daily_int = xclim.indices.daily_pr_intensity(pr, thresh='5 mm/day', freq="QS-DEC")
     """
-    t = utils.convert_units_to(thresh, pr, "hydro")
+    t = convert_units_to(thresh, pr, "hydro")
 
     # put pr=0 for non wet-days
     pr_wd = xarray.where(pr >= t, pr, 0)
@@ -134,7 +135,7 @@ def daily_pr_intensity(pr, thresh="1 mm/day", freq="YS"):
 
     # sum over wanted period
     s = pr_wd.resample(time=freq).sum(dim="time", keep_attrs=True)
-    sd = utils.pint_multiply(s, 1 * units.day, "mm")
+    sd = pint_multiply(s, 1 * units.day, "mm")
 
     # get number of wetdays over period
     wd = wetdays(pr, thresh=thresh, freq=freq)
@@ -178,7 +179,7 @@ def maximum_consecutive_wet_days(
     where :math:`[P]` is 1 if :math:`P` is true, and 0 if false. Note that this formula does not handle sequences at
     the start and end of the series, but the numerical algorithm does.
     """
-    thresh = utils.convert_units_to(thresh, pr, "hydro")
+    thresh = convert_units_to(thresh, pr, "hydro")
 
     group = (pr > thresh).resample(time=freq)
     return group.apply(rl.longest_run, dim="time")
@@ -217,7 +218,7 @@ def cooling_degree_days(
 
     where :math:`[P]` is 1 if :math:`P` is true, and 0 if false.
     """
-    thresh = utils.convert_units_to(thresh, tas)
+    thresh = convert_units_to(thresh, tas)
 
     return (
         tas.pipe(lambda x: x - thresh).clip(min=0).resample(time=freq).sum(dim="time")
@@ -262,7 +263,7 @@ def freshet_start(
     is true, where :math:`w` is the number of days the temperature threshold should be exceeded,  and :math:`[P]` is
     1 if :math:`P` is true, and 0 if false.
     """
-    thresh = utils.convert_units_to(thresh, tas)
+    thresh = convert_units_to(thresh, tas)
     over = tas > thresh
     return over.resample(time=freq).map(
         rl.first_run, dim="time", window=window, coord="dayofyear"
@@ -300,7 +301,7 @@ def growing_degree_days(
 
         GD4_j = \sum_{i=1}^I (TG_{ij}-{4} | TG_{ij} > {4}℃)
     """
-    thresh = utils.convert_units_to(thresh, tas)
+    thresh = convert_units_to(thresh, tas)
     return (
         tas.pipe(lambda x: x - thresh).clip(min=0).resample(time=freq).sum(dim="time")
     )
@@ -363,7 +364,7 @@ def growing_season_length(
 
     >>> gsl = growing_season_length(tas, mid_date='01-01', freq='AS-Jul')
     """
-    thresh = utils.convert_units_to(thresh, tas)
+    thresh = convert_units_to(thresh, tas)
 
     mid_doy = datetime.datetime.strptime(mid_date, "%m-%d").timetuple().tm_yday
 
@@ -426,7 +427,7 @@ def heat_wave_index(
     DataArray
       Heat wave index.
     """
-    thresh = utils.convert_units_to(thresh, tasmax)
+    thresh = convert_units_to(thresh, tasmax)
     over = tasmax > thresh
     group = over.resample(time=freq)
 
@@ -464,7 +465,7 @@ def heating_degree_days(
 
         HD17_j = \sum_{i=1}^{I} (17℃ - TG_{ij})
     """
-    thresh = utils.convert_units_to(thresh, tas)
+    thresh = convert_units_to(thresh, tas)
 
     return tas.pipe(lambda x: thresh - x).clip(0).resample(time=freq).sum(dim="time")
 
@@ -500,8 +501,8 @@ def tn_days_below(
 
         TX_{ij} < Threshold [℃]
     """
-    thresh = utils.convert_units_to(thresh, tasmin)
-    f1 = utils.threshold_count(tasmin, "<", thresh, freq)
+    thresh = convert_units_to(thresh, tasmin)
+    f1 = threshold_count(tasmin, "<", thresh, freq)
     return f1
 
 
@@ -536,7 +537,7 @@ def tx_days_above(
 
         TX_{ij} > Threshold [℃]
     """
-    thresh = utils.convert_units_to(thresh, tasmax)
+    thresh = convert_units_to(thresh, tasmax)
     f = (tasmax > thresh) * 1
     return f.resample(time=freq).sum(dim="time")
 
@@ -572,7 +573,7 @@ def warm_day_frequency(
         TN_{ij} > Threshold [℃]
 
     """
-    thresh = utils.convert_units_to(thresh, tasmax)
+    thresh = convert_units_to(thresh, tasmax)
     events = (tasmax > thresh) * 1
     return events.resample(time=freq).sum(dim="time")
 
@@ -599,7 +600,7 @@ def warm_night_frequency(
     xarray.DataArray
       The number of days with tasmin > thresh per period
     """
-    thresh = utils.convert_units_to(thresh, tasmin)
+    thresh = convert_units_to(thresh, tasmin)
     events = (tasmin > thresh) * 1
     return events.resample(time=freq).sum(dim="time")
 
@@ -635,7 +636,7 @@ def wetdays(pr: xarray.DataArray, thresh: str = "1.0 mm/day", freq: str = "YS"):
     >>> pr = xr.open_dataset('pr.day.nc').pr
     >>> wd = xclim.indices.wetdays(pr, pr_min=5., freq="QS-DEC")
     """
-    thresh = utils.convert_units_to(thresh, pr, "hydro")
+    thresh = convert_units_to(thresh, pr, "hydro")
 
     wd = (pr >= thresh) * 1
     return wd.resample(time=freq).sum(dim="time")
@@ -678,7 +679,7 @@ def maximum_consecutive_dry_days(
     where :math:`[P]` is 1 if :math:`P` is true, and 0 if false. Note that this formula does not handle sequences at
     the start and end of the series, but the numerical algorithm does.
     """
-    t = utils.convert_units_to(thresh, pr, "hydro")
+    t = convert_units_to(thresh, pr, "hydro")
     group = (pr < t).resample(time=freq)
 
     return group.apply(rl.longest_run, dim="time")
@@ -720,7 +721,7 @@ def maximum_consecutive_tx_days(
     where :math:`[P]` is 1 if :math:`P` is true, and 0 if false. Note that this formula does not handle sequences at
     the start and end of the series, but the numerical algorithm does.
     """
-    t = utils.convert_units_to(thresh, tasmax)
+    t = convert_units_to(thresh, tasmax)
     group = (tasmax > t).resample(time=freq)
 
     return group.apply(rl.longest_run, dim="time")
@@ -756,8 +757,8 @@ def sea_ice_area(sic, area, thresh="15 pct"):
     <https://nsidc.org/arcticseaicenews/faq/#area_extent>`_
 
     """
-    t = utils.convert_units_to(thresh, sic)
-    factor = utils.convert_units_to("100 pct", sic)
+    t = convert_units_to(thresh, sic)
+    factor = convert_units_to("100 pct", sic)
     out = xarray.dot(sic.where(sic >= t, 0), area) / factor
     out.attrs["units"] = area.units
     return out
@@ -792,7 +793,7 @@ def sea_ice_extent(sic, area, thresh="15 pct"):
     `What is the difference between sea ice area and extent
     <https://nsidc.org/arcticseaicenews/faq/#area_extent>`_
     """
-    t = utils.convert_units_to(thresh, sic)
+    t = convert_units_to(thresh, sic)
     out = xarray.dot(sic >= t, area)
     out.attrs["units"] = area.units
     return out
@@ -829,7 +830,7 @@ def tropical_nights(
 
         TN_{ij} > Threshold [℃]
     """
-    thresh = utils.convert_units_to(thresh, tasmin)
+    thresh = convert_units_to(thresh, tasmin)
     return (
         tasmin.pipe(lambda x: (tasmin > thresh) * 1).resample(time=freq).sum(dim="time")
     )
