+++ xclim/indices/generic.py
@@ -1,5 +1,11 @@
 # -*- coding: utf-8 -*-
-# Note: stats.dist.shapes: comma separated names of shape parameters
+"""
+Generic indices submodule
+=========================
+
+Helper functions for common generic actions done in the computation of indices.
+"""
+# Note: scipy.stats.dist.shapes: comma separated names of shape parameters
 # The other parameters, common to all distribution, are loc and scale.
 from typing import Sequence
 from typing import Union
@@ -303,3 +309,138 @@ def get_dist(dist):
         e = f"Statistical distribution `{dist}` is not found in scipy.stats."
         raise ValueError(e)
     return dc
+
+
+binary_ops = {">": "gt", "<": "lt", ">=": "ge", "<=": "le"}
+
+
+def threshold_count(
+    da: xr.DataArray, op: str, thresh: float, freq: str
+) -> xr.DataArray:
+    """Count number of days above or below threshold.
+
+    Parameters
+    ----------
+    da : xr.DataArray
+      Input data.
+    op : str
+      Logical operator {>, <, >=, <=, gt, lt, ge, le }. e.g. arr > thresh.
+    thresh : float
+      Threshold value.
+    freq : str
+      Resampling frequency defining the periods
+      defined in http://pandas.pydata.org/pandas-docs/stable/timeseries.html#resampling.
+
+    Returns
+    -------
+    xr.DataArray
+      The number of days meeting the constraints for each period.
+    """
+    from xarray.core.ops import get_op
+
+    if op in binary_ops:
+        op = binary_ops[op]
+    elif op in binary_ops.values():
+        pass
+    else:
+        raise ValueError(f"Operation `{op}` not recognized.")
+
+    func = getattr(da, "_binary_op")(get_op(op))
+    c = func(da, thresh) * 1
+    return c.resample(time=freq).sum(dim="time")
+
+
+def get_daily_events(da: xr.DataArray, da_value: float, operator: str) -> xr.DataArray:
+    r"""
+    function that returns a 0/1 mask when a condition is True or False
+
+    the function returns 1 where operator(da, da_value) is True
+                         0 where operator(da, da_value) is False
+                         nan where da is nan
+
+    Parameters
+    ----------
+    da : xr.DataArray
+    da_value : float
+    operator : str
+
+
+    Returns
+    -------
+    xr.DataArray
+
+    """
+    events = operator(da, da_value) * 1
+    events = events.where(~(np.isnan(da)))
+    events = events.rename("events")
+    return events
+
+
+def daily_downsampler(da: xr.DataArray, freq: str = "YS") -> xr.DataArray:
+    r"""Daily climate data downsampler
+
+    Parameters
+    ----------
+    da : xr.DataArray
+    freq : str
+
+    Returns
+    -------
+    xr.DataArray
+
+    Note
+    ----
+
+        Usage Example
+
+            grouper = daily_downsampler(da_std, freq='YS')
+            x2 = grouper.mean()
+
+            # add time coords to x2 and change dimension tags to time
+            time1 = daily_downsampler(da_std.time, freq=freq).first()
+            x2.coords['time'] = ('tags', time1.values)
+            x2 = x2.swap_dims({'tags': 'time'})
+            x2 = x2.sortby('time')
+    """
+
+    # generate tags from da.time and freq
+    if isinstance(da.time.values[0], np.datetime64):
+        years = [f"{y:04d}" for y in da.time.dt.year.values]
+        months = [f"{m:02d}" for m in da.time.dt.month.values]
+    else:
+        # cannot use year, month, season attributes, not available for all calendars ...
+        years = [f"{v.year:04d}" for v in da.time.values]
+        months = [f"{v.month:02d}" for v in da.time.values]
+    seasons = [
+        "DJF DJF MAM MAM MAM JJA JJA JJA SON SON SON DJF".split()[int(m) - 1]
+        for m in months
+    ]
+
+    n_t = da.time.size
+    if freq == "YS":
+        # year start frequency
+        l_tags = years
+    elif freq == "MS":
+        # month start frequency
+        l_tags = [years[i] + months[i] for i in range(n_t)]
+    elif freq == "QS-DEC":
+        # DJF, MAM, JJA, SON seasons
+        # construct tags from list of season+year, increasing year for December
+        ys = []
+        for i in range(n_t):
+            m = months[i]
+            s = seasons[i]
+            y = years[i]
+            if m == "12":
+                y = str(int(y) + 1)
+            ys.append(y + s)
+        l_tags = ys
+    else:
+        raise RuntimeError(f"Frequency `{freq}` not implemented.")
+
+    # add tags to buffer DataArray
+    buffer = da.copy()
+    buffer.coords["tags"] = ("time", l_tags)
+
+    # return groupby according to tags
+    return buffer.groupby("tags")
