--- xclim/stats.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Statistical distribution fit module
-"""
